requirements.txt
python-dotenv==0.21.0
selenium==3.8.0
termcolor==1.1.0
